requirements.txt
Django==1.8.4
Markdown==2.6.2
django-filter==0.11.0
django-extensions==1.6.1
virtualenv==13.1.0
psycopg2==2.4.5
django-cors-headers==1.1.0
django-rest-swagger==0.3.4
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
drf-nested-routers==0.10.0
gunicorn==19.3.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
djrill==2.0
Pillow==2.3.0
django-appconf==1.0.2
django-imagekit==3.3
pilkit==1.1.13